Move add VAT details card to dataviews as button
