+++ client/me/purchases/billing-history/billing-history-list-data-view.tsx
@@ -1,7 +1,13 @@
+import config from '@automattic/calypso-config';
 import { Gridicon } from '@automattic/components';
+import { Button } from '@wordpress/components';
 import { DataViews, View } from '@wordpress/dataviews';
 import { useTranslate } from 'i18n-calypso';
 import { useSelector } from 'react-redux';
+import { useGeoLocationQuery } from 'calypso/data/geo/use-geolocation-query';
+import { vatDetails as vatDetailsPath } from 'calypso/me/purchases/paths';
+import useVatDetails from 'calypso/me/purchases/vat-info/use-vat-details';
+import { useTaxName } from 'calypso/my-sites/checkout/src/hooks/use-country-list';
 import getPastBillingTransactions from 'calypso/state/selectors/get-past-billing-transactions';
 import isRequestingBillingTransactions from 'calypso/state/selectors/is-requesting-billing-transactions';
 import { usePagination } from '../use-pagination';
@@ -11,7 +17,6 @@ import { useTransactionsFiltering } from './hooks/use-transactions-filtering';
 import { useTransactionsSorting } from './hooks/use-transactions-sorting';
 import { useViewStateUpdate } from './hooks/use-view-state-update';
 import type { ViewStateUpdate } from './data-views-types';
-
 import 'calypso/components/dataviews/style.scss';
 import './style-data-view.scss';
 
@@ -31,6 +36,26 @@ export default function BillingHistoryListDataView( {
 	const viewState = useViewStateUpdate();
 	const receiptActions = useReceiptActions( getReceiptUrlFor );
 
+	const translate = useTranslate();
+	const { vatDetails } = useVatDetails();
+	const { data: geoData } = useGeoLocationQuery();
+	const taxName = useTaxName( vatDetails.country ?? geoData?.country_short ?? 'GB' );
+
+	const genericTaxName =
+		/* translators: This is a generic name for taxes to use when we do not know the user's country. */
+		translate( 'tax (VAT/GST/CT)' );
+	const fallbackTaxName = genericTaxName;
+	/* translators: %s is the name of taxes in the country (eg: "VAT" or "GST"). */
+	const editVatText = translate( 'Edit %s details', {
+		textOnly: true,
+		args: [ taxName ?? fallbackTaxName ],
+	} );
+	/* translators: %s is the name of taxes in the country (eg: "VAT" or "GST"). */
+	const addVatText = translate( 'Add %s details', {
+		textOnly: true,
+		args: [ taxName ?? fallbackTaxName ],
+	} );
+	const vatText = vatDetails.id ? editVatText : addVatText;
 	const actions = receiptActions.map( ( action ) => ( {
 		...action,
 		icon: <Gridicon icon={ action.iconName } />,
@@ -44,7 +69,7 @@ export default function BillingHistoryListDataView( {
 		viewState.view.page,
 		viewState.view.perPage
 	);
-	const translate = useTranslate();
+
 	const fields = useFieldDefinitions( transactions );
 
 	const handleViewChange = ( view: View ) => viewState.updateView( view as ViewStateUpdate );
@@ -52,6 +77,13 @@ export default function BillingHistoryListDataView( {
 	return (
 		<DataViews
 			data={ paginatedItems }
+			header={
+				config.isEnabled( 'me/vat-details' ) && (
+					<Button className="dataviews__tax-details-notice" variant="link" href={ vatDetailsPath }>
+						{ vatText }
+					</Button>
+				)
+			}
 			paginationInfo={ {
 				totalItems,
 				totalPages,
