+++ client/me/purchases/billing-history/main.tsx
@@ -1,3 +1,4 @@
+import config from '@automattic/calypso-config';
 import { CompactCard, Card } from '@automattic/components';
 import { useTranslate } from 'i18n-calypso';
 import DocumentHead from 'calypso/components/data/document-head';
@@ -5,14 +6,11 @@ import QueryBillingTransactions from 'calypso/components/data/query-billing-tran
 import InlineSupportLink from 'calypso/components/inline-support-link';
 import Main from 'calypso/components/main';
 import NavigationHeader from 'calypso/components/navigation-header';
-import { useGeoLocationQuery } from 'calypso/data/geo/use-geolocation-query';
 import PageViewTracker from 'calypso/lib/analytics/page-view-tracker';
 import BillingHistoryListDataView from 'calypso/me/purchases/billing-history/billing-history-list-data-view';
-import { vatDetails as vatDetailsPath, billingHistoryReceipt } from 'calypso/me/purchases/paths';
+import { billingHistoryReceipt } from 'calypso/me/purchases/paths';
 import PurchasesNavigation from 'calypso/me/purchases/purchases-navigation';
 import titles from 'calypso/me/purchases/titles';
-import useVatDetails from 'calypso/me/purchases/vat-info/use-vat-details';
-import { useTaxName } from 'calypso/my-sites/checkout/src/hooks/use-country-list';
 
 import './style.scss';
 
@@ -32,25 +30,6 @@ export function BillingHistoryContent( {
 
 function BillingHistory() {
 	const translate = useTranslate();
-	const { vatDetails } = useVatDetails();
-	const { data: geoData } = useGeoLocationQuery();
-	const taxName = useTaxName( vatDetails.country ?? geoData?.country_short ?? 'GB' );
-
-	const genericTaxName =
-		/* translators: This is a generic name for taxes to use when we do not know the user's country. */
-		translate( 'tax (VAT/GST/CT)' );
-	const fallbackTaxName = genericTaxName;
-	/* translators: %s is the name of taxes in the country (eg: "VAT" or "GST"). */
-	const editVatText = translate( 'Edit %s details', {
-		textOnly: true,
-		args: [ taxName ?? fallbackTaxName ],
-	} );
-	/* translators: %s is the name of taxes in the country (eg: "VAT" or "GST"). */
-	const addVatText = translate( 'Add %s details', {
-		textOnly: true,
-		args: [ taxName ?? fallbackTaxName ],
-	} );
-	const vatText = vatDetails.id ? editVatText : addVatText;
 
 	return (
 		<Main id="purchases" wideLayout>
