+++ client/me/purchases/billing-history/style-data-view.scss
@@ -85,7 +85,7 @@
 					font-weight: 600;
 					margin-bottom: 0.25rem;
 				}
-	
+
 				small {
                     display: block;
                     font-size: 0.875rem;
@@ -103,7 +103,7 @@
             td:last-child {
                 padding-right: 24px;
             }
-			
+
 			.dataviews-view-table__actions-column {
 				text-align: right;
 			}
@@ -182,11 +182,45 @@
         }
     }
 
-    .dataviews__view-actions {
-        margin-top: 0;
-		padding-left: 24px;
-		padding-right: 24px;
+    .dataviews__search .dataviews-search {
+        min-width: 160px;
+
+        & .components-input-control__input {
+            font-size: 0.875rem; /* 13px equivalent */;
+        }
     }
+
+    .dataviews__view-actions {
+        display: flex;
+        flex-wrap: wrap;
+        justify-content: space-between;
+        margin-top: 1em;
+		padding-left: 0;
+		padding-right: 0;
+        row-gap: .5rem;
+
+        & .dataviews__tax-details-notice {
+            text-decoration: none;
+            order: -1;
+        }
+      }
+
+      .dataviews__view-actions > * {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+        width: 100%; /* Default full width on narrow screens */
+      }
+
+      @media (min-width: 600px) {
+        .dataviews__view-actions {
+          flex-wrap: nowrap;
+        }
+
+        .dataviews__view-actions > * {
+          width: auto; /* Auto width when space allows */
+        }
+      }
 
 	.dataviews-filters__container {
 		padding: 0;
